pyproject.toml
[tool.poetry]
name = "http_pubsub_tools"
version = "0.1.0"
description = "Easily process PubSub messages from HTTP handlers with batching control."
authors = ["gmega <giuliano.mega@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-pubsub = "2.9.0"
avro = "1.11.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mypy = "^0.971"
google-api-python-client-stubs = "^1.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = [
    "avro",
    "google.cloud"
]
ignore_missing_imports = "True"